+++ src/pages/api/users/[id].ts
@@ -4,6 +4,7 @@ import { hash } from "bcrypt";
 import User, {
   userInputServerValidator,
 } from "../../../../server/mongodb/models/User";
+import { isAdmin, isOwnUser, isOriginalOrgAdmin } from "../../../utils/routeProtection";
 
 export default async (req: NextApiRequest, res: NextApiResponse) => {
   await dbConnect();
@@ -18,6 +19,21 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
       return res.status(200).json({ user });
     }
     case "PUT": {
+      const isadmin = await isAdmin(req, res);
+      const isownuser = await isOwnUser(req, res);
+      const isogadmin = await isOriginalOrgAdmin(req, res);
+      if (user?.role === "admin" && !isogadmin) {
+        return res
+        .status(403)
+        .json({ error: "Only the original organization Admin can modify other admins" });
+      }
+
+      if (!isadmin && !isownuser) {
+        return res
+        .status(403)
+        .json({ error: "Only Admins can modify other users, and volunteers can only modify themselves" });
+      }
+
       const result = userInputServerValidator.partial().safeParse(req.body);
       if (!result.success) return res.status(400).json(result);
 
@@ -32,6 +48,21 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
       return res.status(200).json({ user });
     }
     case "POST": {
+      const isadmin = await isAdmin(req, res);
+      const isownuser = await isOwnUser(req, res);
+      const isogadmin = await isOriginalOrgAdmin(req, res);
+      if (user?.role === "admin" && !isogadmin) {
+        return res
+        .status(403)
+        .json({ error: "Only the origianl organization Admin can modify other admins" });
+      }
+      // only admins can modify other users, and volunteers can only modify themselves
+      if (!isadmin && !isownuser) {
+        return res
+        .status(403)
+        .json({ error: "Only Admins can modify other users, and volunteers can only modify themselves" });
+      }
+
       const result = userInputServerValidator.partial().safeParse(req.body);
       if (!result.success) return res.status(400).json(result);
 
@@ -45,6 +76,20 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
       return res.status(200).json({ success: true, user });
     }
     case "DELETE": {
+      // TODO: should admins be able to delete other admins?
+      const isadmin = await isAdmin(req, res);
+      const isogadmin = await isOriginalOrgAdmin(req, res);
+
+      if (user?.role === "admin" && !isogadmin) {
+        return res
+        .status(403)
+        .json({ error: "Only the original organization Admin can delete other admins" });
+      }
+      if (!isadmin) {
+        return res
+          .status(403)
+          .json({ error: "Only Admins can delete users" });
+      }
       await user.deleteOne();
       return res.status(204).end();
     }
